+++ dc-model/src/main/java/de/digitalcollections/model/identifiable/entity/Project.java
@@ -44,7 +44,6 @@ public class Project extends Entity {
   @Override
   protected void init() {
     super.init();
-    this.entityType = EntityType.PROJECT;
   }
 
   /**